extract DownloadFeaturette
